Revert "Revert "renamed init""
This reverts commit b042a7a8bacd9368617dd6b1321b275c4a448076.
