+++ DashboardProject/views.py
@@ -7,3 +7,4 @@ views = Blueprint('views', __name__)
 @views.route('/')
 def home():
     return render_template("index.html")
+
